huge strongly connected components (#525)
* migrate tarjan to huge

* WIP

* hugified scc-tarjan and iterative-tarjan

* add testcase & iface for abstraction from huge+

* impl. writeback, tests, refactorinh

* WIP

* WIP

+++ algo/src/main/java/org/neo4j/graphalgo/results/SCCStreamResult.java
@@ -34,7 +34,7 @@ public class SCCStreamResult {
      */
     public final long partition;
 
-    public SCCStreamResult(long nodeId, int clusterId) {
+    public SCCStreamResult(long nodeId, long clusterId) {
         this.nodeId = nodeId;
         this.partition = clusterId;
     }
